refactor(webview): split high-impact fallow files
Split the three fallow split_high_impact targets into focused modules,
behavior-preserving (no exported signature or behavior change):

- memoryData: extract integrityHighlight.ts (paint concern)
- SearchBar: extract searchBarRender.ts (pure helpers + searchKeyFor/
  SearchTrigger); searchEngine/test import from it
- HexView: extract HexViewRender.ts (pure DOM-free render + types) and
  HexViewPaint.ts (DOM paint/match utilities); grid/test/viewer import
  from the owning module

Class modules keep only the interaction class, dropping fan-in to class
consumers. Fallow now reports 0 issues / 0 findings / 0 dupes / 0 targets.
