+++ .trellis/tasks/08-05-fallow-split-targets/task.json
@@ -3,7 +3,7 @@
   "name": "fallow-split-targets",
   "title": "Split high-impact files flagged by fallow",
   "description": "Split the three fallow health.targets split_high_impact files: memoryData.ts (38 LOC/11 deps), SearchBar.ts (260 LOC), HexView.ts (627 LOC). Behavior-preserving splits; reduce per-change blast radius.",
-  "status": "planning",
+  "status": "in_progress",
   "dev_type": null,
   "scope": null,
   "package": null,
