+++ src/webview/components/SearchBar/SearchBar.ts
@@ -6,11 +6,9 @@
 // `S` global and never calls engine functions directly.
 
 import { esc } from '../../utils';
-import { canonicalizeQuery } from '../../../core/search';
+import { activeClass, modeOptions, PLACEHOLDERS, type SearchTrigger } from './searchBarRender';
 import type { SearchEndianness, SearchMode } from '../../../core/types';
 import './SearchBar.css';
-
-export type SearchTrigger = 'enter-next' | 'enter-prev' | 'button';
 
 export interface SearchBarCallbacks {
     onSearch: (query: string, mode: SearchMode, endianness: SearchEndianness, trigger: SearchTrigger) => void;
@@ -23,37 +21,6 @@ export interface SearchBarSeedOptions {
     mode?: SearchMode;
     endianness?: SearchEndianness;
     query?: string;
-}
-
-const MODE_LABELS: ReadonlyArray<[SearchMode, string]> = [
-    ['bytes', 'Bytes'],
-    ['value', 'Value'],
-    ['ascii', 'ASCII'],
-    ['addr', 'Addr'],
-];
-
-const PLACEHOLDERS: Record<SearchMode, string> = {
-    bytes: 'Bytes (e.g. DE AD BE EF)',
-    value: 'Value (e.g. 0x12345678 or 305419896)',
-    ascii: 'ASCII text',
-    addr: 'Addr (e.g. 1A0)',
-};
-
-/** Canonical search key — engine reuses it for running-search parity. */
-export function searchKeyFor(mode: SearchMode, raw: string, endianness: SearchEndianness): string {
-    const canonical = canonicalizeQuery(mode, raw);
-    const endianKey = mode === 'value' ? endianness : 'n/a';
-    return `${mode}|${endianKey}|${canonical}`;
-}
-
-function activeClass(active: boolean): string {
-    return active ? 'active' : '';
-}
-
-function modeOptions(selected: SearchMode): string {
-    return MODE_LABELS
-        .map(([value, label]) => `<option value="${value}"${value === selected ? ' selected' : ''}>${label}</option>`)
-        .join('');
 }
 
 export class SearchBar {
